requirements.txt
Django~=3.2.13
gunicorn~=20.1.0
psycopg2-binary~=2.9
graphene-django==2.15.0
django-graphql-auth~=0.3.16
